chore(dataplanes): move envoy related sources to connections
