+++ packages/kuma-gui/src/app/connections/views/ConnectionOutboundSummaryClustersView.vue
@@ -60,5 +60,5 @@
   </RouteView>
 </template>
 <script lang="ts" setup>
-import { sources } from '@/app/data-planes/sources'
+import { sources } from '../sources'
 </script>
