+++ packages/kuma-gui/src/app/connections/views/ConnectionOutboundSummaryStatsView.vue
@@ -9,7 +9,7 @@
       connection: '',
     }"
     name="connection-outbound-summary-stats-view"
-    v-slot="{ route }"
+    v-slot="{ route, uri }"
   >
     <RouteTitle
       :render="false"
@@ -17,8 +17,13 @@
     />
     <AppView>
       <DataLoader
-        :src="`/meshes/${route.params.mesh}/dataplanes/${route.params.dataPlane}/stats/${props.dataplaneOverview.dataplane.networking.inboundAddress}`"
-        v-slot="{ data, refresh }: StatsSource"
+        :src="uri(sources, '/meshes/:mesh/dataplanes/:name/stats/:address', {
+          mesh: route.params.mesh,
+          name: route.params.dataPlane,
+          address: props.dataplaneOverview.dataplane.networking.inboundAddress,
+        })"
+
+        v-slot="{ data, refresh }"
       >
         <DataCollection
           :items="data!.raw.split('\n')"
@@ -52,7 +57,7 @@
   </RouteView>
 </template>
 <script lang="ts" setup>
-import { StatsSource } from '../sources'
+import { sources } from '../sources'
 import type { DataplaneOverview } from '@/app/data-planes/data/'
 const props = defineProps<{
   dataplaneOverview: DataplaneOverview
