+++ packages/kuma-gui/src/app/data-planes/views/DataPlaneXdsConfigView.vue
@@ -57,5 +57,5 @@
 </template>
 
 <script lang="ts" setup>
-import { sources } from '../sources'
+import { sources } from '@/app/connections/sources'
 </script>
